ticket_service/requirements.txt
celery
redis==4.5.1
Flask==2.2.5
opentelemetry-api==1.32.0
opentelemetry-sdk==1.32.0
opentelemetry-exporter-otlp-proto-http==1.32.0
opentelemetry-instrumentation==0.53b0
opentelemetry-instrumentation-redis==0.53b0
opentelemetry-semantic-conventions==0.53b0